refactor(public, src): Split code into modules
